pyproject.toml
[project]
name = "nextmeeting"
version = "2.0.0"
description = "Show your nextmeeting in your poly/waybar with gcalcli"
readme = "README.md"
authors = [{ name = "Chmouel Boudjnah", email = "chmouel@chmouel.com" }]
requires-python = ">=3.9"
license = "Apache-2.0"
dependencies = ["python-dateutil>=2.9.0.post0"]
classifiers = ["License :: OSI Approved :: Apache Software License"]
keywords = ["calendar", "cli"]

[project.scripts]
nextmeeting = "nextmeeting.cli:main"
nextmeetingd = "nextmeeting.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
ignore = ["E501", "PLR0912", "PLR0915"]
select = ["E", "F", "D4", "PT", "PL"]

[tool.pylint]
disable = [
  "missing-module-docstring",
  "missing-class-docstring",
  "missing-function-docstring",
  "broad-exception-caught",
  "invalid-name",
  "too-many-branches",
]
max-line-length = 101

[dependency-groups]
dev = [
  "isort>=6.0.1",
  "pytest>=8.4.1",
  "pytest-cov>=6.2.1",
  "types-python-dateutil>=2.9.0.20250516",
]
